requirements.txt
langchain_community
langchain_core
langchain_openai

faiss-cpu
tqdm
python-dotenv

gradio